save data and generate html for payment summary
