+++ data/deliveryOptions.js
@@ -13,3 +13,15 @@ export const deliveryOptions = [
     priceCents : 999
   },
 ]
+
+export function getDeliveryOption(deliveryOptionsId) {
+  let deliveryOption;
+
+    deliveryOptions.forEach((option) => {
+      if (option.id === deliveryOptionsId) {
+        deliveryOption = option;
+      }
+    });
+
+    return deliveryOption || deliveryOption[0];
+}
