+++ scripts/checkout.js
@@ -1,158 +1,5 @@
-import { cart, removeFromCart, updateDeliveryDate } from "../data/cart.js";
-import { products } from "../data/products.js";
-import { formatCurrancy } from "./utils/money.js";
-// import {hello} from 'https://unpkg.com/supersimpledev@1.0.1/hello.esm.js';
-import { deliveryOptions } from "../data/deliveryOptions.js";
-
-
-// hello();
-
-const today = dayjs();
-const deliveryDate = today.add(7, 'day');
-
-console.log(deliveryDate.format('dddd, MMMM D'));
-
-function renderOrderSummary() {
-  
-
-    let cartSummryHtml = '';
-
-    cart.forEach((cartItem) => {
-    const productId = cartItem.productId;
-    let matchingProduct;
-    products.forEach((product) => {
-      if (product.id === productId) {
-        matchingProduct = product;
-      }
-    }
-    );
-
-
-    const deliveryOptionsId = cartItem.deliveryOptionsId;
-    let deliveryOption;
-
-    deliveryOptions.forEach((option) => {
-      if (option.id === deliveryOptionsId) {
-        deliveryOption = option;
-      }
-    });
-
-    const today = dayjs();
-        const deliveryDate = today.add(
-          deliveryOption.deliveryDays, 'days'
-        );
-        const dateString = deliveryDate.format('dddd, MMMM D');
-
-    cartSummryHtml += `
-    <div class="cart-item-container js-cart-item-container-${matchingProduct.id}">
-        <div class="delivery-date">
-          Delivery date: ${dateString}
-        </div>
-
-        <div class="cart-item-details-grid">
-          <img class="product-image"
-            src="${matchingProduct.image}">
-
-          <div class="cart-item-details">
-            <div class="product-name">
-              ${matchingProduct.name}
-            </div>
-            <div class="product-price">
-              $${formatCurrancy(matchingProduct.priceCents)}
-            </div>
-            <div class="product-quantity">
-              <span>
-                Quantity: <span class="quantity-label">${cartItem.quantity}</span>
-              </span>
-              <span class="update-quantity-link link-primary">
-                Update
-              </span>
-              <span class="delete-quantity-link link-primary js-delete-link" data-product-id="${matchingProduct.id}">
-                Delete
-              </span>
-            </div>
-          </div>
-
-          <div class="delivery-options">
-            <div class="delivery-options-title">
-              Choose a delivery option:
-            </div>
-            
-            ${deliveryOptionsHtml(matchingProduct,cartItem)}
-          </div>
-        </div>
-      </div>
-    `
-    }
-    );
-
-    function deliveryOptionsHtml(matchingProduct, cartItem) {
-    let html ='';
-    deliveryOptions.forEach(
-      (deliveryOption) => {
-
-        const today = dayjs();
-        const deliveryDate = today.add(
-          deliveryOption.deliveryDays, 'days'
-        );
-        const dateString = deliveryDate.format('dddd, MMMM D');
-        const priceStirng = deliveryOption.priceCents ===0
-                            ? 'FREE'
-                            : `$${formatCurrancy(deliveryOption.priceCents)} -` ;
-
-        const isChecked = deliveryOption.id === cartItem.deliveryOptionsId;
-
-        html +=`<div class="delivery-option js-delivery-option"
-                  data-product-id = "${matchingProduct.id}"
-                  data-delivery-option-id = "${deliveryOption.id}" >
-              <input type="radio" 
-                ${ isChecked ? 'checked': ''}
-                class="delivery-option-input"
-                name="delivery-option-${matchingProduct.id}">
-              <div>
-                <div class="delivery-option-date">
-                  ${dateString}
-                </div>
-                <div class="delivery-option-price">
-                  ${priceStirng} Shipping
-                </div>
-              </div>
-          </div> 
-          `
-      }
-    )
-    return html;
-    }
-
-    document.querySelector('.js-order-summary').innerHTML = cartSummryHtml;
-
-    document.querySelectorAll('.js-delete-link')
-    .forEach((link) => {
-      link.addEventListener('click',
-        () => {
-          const productId = link.dataset.productId;
-
-          removeFromCart(productId);
-
-          const container = document.querySelector(`.js-cart-item-container-${productId}`);
-
-          container.remove();
-        }
-      );
-
-    }
-    )
-
-    document.querySelectorAll('.js-delivery-option').forEach((element) => {
-      element.addEventListener( 'click', ()=>{
-        const productId = element.dataset.productId ;
-        const deliveryOptionId = element.dataset.deliveryOptionId ;
-
-        updateDeliveryDate(productId, deliveryOptionId);
-        renderOrderSummary();
-      });
-    });
-
-}
+import { renderOrderSummary } from "./checkout/orderSummary.js";
+import { renderPaymetSummary } from "./checkout/paymentSummary.js";
 
 renderOrderSummary();
+renderPaymetSummary();
