+++ scripts/checkout/paymentSummary.js
@@ -1,0 +1,62 @@
+import { cart } from "../../data/cart.js";
+import { getProduct } from "../../data/products.js";
+import { deliveryOptions, getDeliveryOption } from "../../data/deliveryOptions.js";
+import { formatCurrancy } from "../utils/money.js";
+
+export function renderPaymetSummary() {
+  let productPriceCents = 0;
+  let shippingPriceCents = 0;
+
+  cart.forEach((cartItem)=>{
+
+    const product = getProduct(cartItem.productId);
+
+    productPriceCents += product.priceCents * cartItem.quantity;
+
+    const deliveryOption= getDeliveryOption(cartItem.deliveryOptionsId);
+    shippingPriceCents += deliveryOption.priceCents;
+  });
+  console.log(productPriceCents);
+  console.log(shippingPriceCents);
+
+  const totalBeforTaxcents = productPriceCents + shippingPriceCents ;
+  const taxCents = totalBeforTaxcents*0.1 ;
+  const totalCents = totalBeforTaxcents + taxCents ;
+
+  const paymentSummaryHtml = 
+  `<div class="payment-summary-title">
+      Order Summary
+    </div>
+
+    <div class="payment-summary-row">
+      <div>Items (3):</div>
+      <div class="payment-summary-money">$${formatCurrancy(productPriceCents)}</div>
+    </div>
+
+    <div class="payment-summary-row">
+      <div>Shipping &amp; handling:</div>
+      <div class="payment-summary-money">$${formatCurrancy(shippingPriceCents)}</div>
+    </div>
+
+    <div class="payment-summary-row subtotal-row">
+      <div>Total before tax:</div>
+      <div class="payment-summary-money">$${formatCurrancy(totalBeforTaxcents)}</div>
+    </div>
+
+    <div class="payment-summary-row">
+      <div>Estimated tax (10%):</div>
+      <div class="payment-summary-money">$${formatCurrancy(taxCents)}</div>
+    </div>
+
+    <div class="payment-summary-row total-row">
+      <div>Order total:</div>
+      <div class="payment-summary-money">$${formatCurrancy(totalCents)}</div>
+    </div>
+
+    <button class="place-order-button button-primary">
+      Place your order
+    </button>
+  `;
+  console.log(paymentSummaryHtml);
+  document.querySelector('.js-payment-summary').innerHTML = paymentSummaryHtml; 
+}
