+++ scripts/utils/money.js
@@ -1,4 +1,4 @@
 
 export function formatCurrancy(priceCents) {
-  return (priceCents/100).toFixed(2);
+  return (Math.round(priceCents)/100).toFixed(2);
 }
